Add files via upload
This is a ML model that I built to help identify leads more likely to convert to paid customers.
Through data analysis I was able to find the key factors driving the lead conversion process and create a profile of the leads which are likely to convert.
I had to deal with unbalanced data and redundant data.  I  engineered features and removed outliers.  I created univariate and bivariate data visualizations, and tried a host of different machine learning models for this prediction program.  The models I used and tested are the Decision Tree Classifier, Bagging Classifier, Random Forest Classifier, AdaBoost Classifier, and Gradient Boosting Classifier.  The tuned Random Forest Classifier performed the best.